+++ NEWS.md
@@ -1,6 +1,9 @@
 27.0 (Not released yet)
 =======================
 
+* `e3-opt-check`: new script to look for syntax errors in opt files.
+* `e3-opt-parse`: exit gracefully and give line number information in case of
+  syntax error.
 * `e3-convert-xunit`: truncate too long test result messages.
 * `e3-convert-xunit`: warn about dangling XFAILs annotations.
 * `DiffTestDriver`: tolerate missing baseline files when rewriting baselines.